Remove the EntryDisplay component and give control of the entries to the react-tables which render them
The EntryDisplay component has evolved outside of its original
purpose. It now merely acts as a middleman between the App component
and the react-tables it renders, aside from controlling the state of
the entries. It seems more expediant to remove the EntryDisplay
component and give control over the displayed entries directly
to the react-tables which render them.